+++ common/home-manager-sets/hyprland-ags-desktop.nix
@@ -2,5 +2,5 @@
 let
   inherit (self.common) home-manager-modules;
   inherit (self.common.home-manager-module-sets) hyprland-desktop;
-in hyprland-desktop ++ (with home-manager-modules; [ ags ])
+in hyprland-desktop ++ (with home-manager-modules; [ ags mopidy ])
 ++ [ self.inputs.ags.homeManagerModules.default ]